trying out using u and v as inputs to the NN rather than derivatives thereof, this matches what was done in other papers and I want to see how the offline problem goes. I'm also comparing the tensors T11, T12, and T22 instead of the parameterizations themselves.

+++ src/run_model.jl
@@ -41,7 +41,7 @@ function run_model(::Type{T},P::Parameter) where {T<:AbstractFloat}
 
 end
 
-function model_setup(::Type{T}=Float64;     # number format
+function model_setup(::Type{T}=Float32;     # number format
     kwargs...                               # all additional parameters
     ) where {T<:AbstractFloat}
